// src/types/opts.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface Opts {
  ARCHIVE_OUTPUT: boolean;
  ARCHIVE_ROOT_FOLDER: boolean;
  CACHE_ACTIVE: boolean;
  EXPORT_DEBUG: boolean;
  EXPORT_PACK_ONLY: boolean;
  PRESETS_TO_EXPORT: string[] | null;
  GODOT_ARCHIVE_PATH: string;
  GODOT_BUILD_PATH: string;
  GODOT_CONFIG_PATH: string;
  GODOT_DOWNLOAD_URL: string;
  GODOT_EXPORT_TEMPLATES_PATH: string;
  GODOT_PROJECT_FILE_PATH: string;
  GODOT_PROJECT_PATH: string;
  GODOT_TEMPLATES_DOWNLOAD_URL: string;
  GODOT_VERBOSE: boolean;
  GODOT_WORKING_PATH: string;
  RELATIVE_EXPORT_PATH: string;
  RELATIVE_PROJECT_PATH: string;
  USE_GODOT_3: boolean;
  USE_PRESET_EXPORT_PATH: boolean;
  WINE_PATH: string;
  EDITOR_SETTINGS_FILENAME: string;
  GODOT_TEMPLATES_PATH: string;
}
/* eslint-enable @typescript-eslint/naming-convention */